+++ arkenv/src/main/kotlin/com/apurebase/arkenv/argument/ArkenvSimpleArgument.kt
@@ -1,6 +1,7 @@
 package com.apurebase.arkenv.argument
 
 import com.apurebase.arkenv.Arkenv
+import com.apurebase.arkenv.module.ArkenvModuleConfiguration
 import kotlin.reflect.KProperty
 
 /**
@@ -16,6 +17,7 @@ internal class ArkenvSimpleArgument<T : Any?>(
     override lateinit var property: KProperty<*>
     override lateinit var arkenv: Arkenv
     override var isSet = false
+    private var isInitialized = false
 
     override val defaultValue: T?
         get() {
@@ -28,8 +30,14 @@ internal class ArkenvSimpleArgument<T : Any?>(
 
     override var value: T = null as T
 
-    override fun initialize(arkenv: Arkenv, property: KProperty<*>) {
+    override fun initialize(arkenv: Arkenv, property: KProperty<*>, moduleConfiguration: ArkenvModuleConfiguration?) {
         this.arkenv = arkenv
         this.property = property
+        if (!isInitialized) {
+            val prefix = moduleConfiguration?.prefix ?: arkenv.configuration.prefix
+            val argumentNameProcessor = ArgumentNameProcessor(prefix)
+            argumentNameProcessor.processArgumentNames(argument, property)
+            isInitialized = true
+        }
     }
 }