+++ arkenv/src/main/kotlin/com/apurebase/arkenv/module/ArkenvModuleConfiguration.kt
@@ -1,0 +1,8 @@
+package com.apurebase.arkenv.module
+
+/**
+ * Module configuration value class.
+ */
+class ArkenvModuleConfiguration(
+    var prefix: String? = null
+)